+++ springboot/src/main/java/ca/jrvs/apps/trading/repository/TraderRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TraderRepository extends JpaRepository<Trader, Long> {
+public interface TraderRepository extends JpaRepository<Trader, Integer> {
 
-  default Optional<Account> findAccountById(Long id) {
+  default Optional<Account> findAccountById(Integer id) {
     if (findById(id).isPresent()) {
       return Optional.of(findById(id).get().getAccount());
     }
